+++ api-e2e/run.js
@@ -1,7 +1,6 @@
 var json = require('./json/merged-tests.json');
 var najax = require('najax');
-//import json from './json/merged-tests.json';
-var { validate, reportValidationResult } = require('./validation');
+var { validate, reportValidationResult, runSpecs } = require('./validation');
 const test = json[1].suites[0].tests[0];
 
 const host = 'http://localhost:8082';
@@ -15,7 +14,15 @@ async function main() {
         test.hash
     );
 
-    console.log(reportValidationResult(result, ''));
+    const fileFilter = 'mutation-data';
+
+    const files = fileFilter?.trim().length
+        ? json.filter(f => new RegExp(fileFilter).test(f.file))
+        : json;
+
+    await runSpecs(files, najax, 'http://localhost:8082');
+
+    //console.log(reportValidationResult(result, ''));
 }
 
 main();
